chore: replace `unwrapKeys` with `excludeParentKeys`
